Created package constraints
Ref #7.
