+++ src/neurom/constraints/__init__.py
@@ -1,0 +1,3 @@
+from neurom.constraints.constraint import Constraint
+from neurom.constraints.no_constraint import NoConstraint
+from neurom.constraints.dirichlet import Dirichlet
